requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
Django==3.2.12
django-rest-swagger==2.2.0
django-separatedvaluesfield==0.6.0
djangorestframework==3.13.1
drf-yasg==1.20.0
filelock==3.6.0
Flask-RESTful==0.3.9
idna==3.3
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
MarkupSafe==2.1.1
platformdirs==2.5.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.13.4
virtualenvwrapper-win==1.2.7
Werkzeug==2.0.3
Pillow==9.2.0